feat: rename auth forms files
